rename ClosureCodegen implicit mentions
