+++ compiler/backend/src/org/jetbrains/kotlin/codegen/CallBasedArgumentGenerator.java
@@ -8,7 +8,7 @@ package org.jetbrains.kotlin.codegen;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.kotlin.descriptors.CallableDescriptor;
 import org.jetbrains.kotlin.descriptors.ValueParameterDescriptor;
-import org.jetbrains.kotlin.implicit.ExtensionCandidate;
+import org.jetbrains.kotlin.extensionresolution.ExtensionCandidate;
 import org.jetbrains.kotlin.psi.KtExpression;
 import org.jetbrains.kotlin.psi.ValueArgument;
 import org.jetbrains.kotlin.resolve.BindingContext;
