+++ compiler/frontend/src/org/jetbrains/kotlin/extensionresolution/ImplicitCandidateResolution.kt
@@ -3,7 +3,7 @@
  * that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.implicit
+package org.jetbrains.kotlin.extensionresolution
 
 sealed class ImplicitCandidateResolution {
     data class Resolved(val candidate: ExtensionCandidate) : ImplicitCandidateResolution()
